engine/rocks: split rocks/mod.rs into multiple mods
