+++ components/engine/src/rocks/mod.rs
@@ -1,13 +1,10 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::fmt::{self, Debug, Formatter};
-use std::ops::Deref;
-use std::option::Option;
-use std::sync::Arc;
-
-use super::{Error, Iterable, Mutable, Peekable, Result};
-
+mod db;
 pub mod util;
+
+mod snapshot;
+pub use self::snapshot::*;
 
 pub use engine_rocksdb::rocksdb_options::UnsafeSnap;
 pub use engine_rocksdb::{
@@ -24,169 +21,11 @@ pub use engine_rocksdb::{
     WriteStallCondition, WriteStallInfo, DB,
 };
 
-use super::IterOption;
-
-pub struct Snapshot {
-    db: Arc<DB>,
-    snap: UnsafeSnap,
-}
-
-/// Because snap will be valid whenever db is valid, so it's safe to send
-/// it around.
-unsafe impl Send for Snapshot {}
-unsafe impl Sync for Snapshot {}
-
-impl Snapshot {
-    pub fn new(db: Arc<DB>) -> Snapshot {
-        unsafe {
-            Snapshot {
-                snap: db.unsafe_snap(),
-                db,
-            }
-        }
-    }
-
-    pub fn into_sync(self) -> SyncSnapshot {
-        SyncSnapshot(Arc::new(self))
-    }
-
-    pub fn cf_names(&self) -> Vec<&str> {
-        self.db.cf_names()
-    }
-
-    pub fn cf_handle(&self, cf: &str) -> Result<&CFHandle> {
-        self::util::get_cf_handle(&self.db, cf).map_err(Error::from)
-    }
-
-    pub fn get_db(&self) -> Arc<DB> {
-        Arc::clone(&self.db)
-    }
-
-    pub fn db_iterator(&self, iter_opt: IterOption) -> DBIterator<Arc<DB>> {
-        let mut opt = iter_opt.build_read_opts();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        DBIterator::new(Arc::clone(&self.db), opt)
-    }
-
-    pub fn db_iterator_cf(&self, cf: &str, iter_opt: IterOption) -> Result<DBIterator<Arc<DB>>> {
-        let handle = self::util::get_cf_handle(&self.db, cf)?;
-        let mut opt = iter_opt.build_read_opts();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        Ok(DBIterator::new_cf(Arc::clone(&self.db), handle, opt))
-    }
-}
-
-impl Debug for Snapshot {
-    fn fmt(&self, fmt: &mut Formatter<'_>) -> fmt::Result {
-        write!(fmt, "Engine Snapshot Impl")
-    }
-}
-
-impl Drop for Snapshot {
-    fn drop(&mut self) {
-        unsafe {
-            self.db.release_snap(&self.snap);
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct SyncSnapshot(Arc<Snapshot>);
-
-impl Deref for SyncSnapshot {
-    type Target = Snapshot;
-
-    fn deref(&self) -> &Snapshot {
-        &self.0
-    }
-}
-
-impl SyncSnapshot {
-    pub fn new(db: Arc<DB>) -> SyncSnapshot {
-        SyncSnapshot(Arc::new(Snapshot::new(db)))
-    }
-
-    pub fn clone(&self) -> SyncSnapshot {
-        SyncSnapshot(Arc::clone(&self.0))
-    }
-}
-
-impl Peekable for DB {
-    fn get_value(&self, key: &[u8]) -> Result<Option<DBVector>> {
-        let v = self.get(key)?;
-        Ok(v)
-    }
-
-    fn get_value_cf(&self, cf: &str, key: &[u8]) -> Result<Option<DBVector>> {
-        let handle = self::util::get_cf_handle(self, cf)?;
-        let v = self.get_cf(handle, key)?;
-        Ok(v)
-    }
-}
-
-impl Iterable for DB {
-    fn new_iterator(&self, iter_opt: IterOption) -> DBIterator<&DB> {
-        self.iter_opt(iter_opt.build_read_opts())
-    }
-
-    fn new_iterator_cf(&self, cf: &str, iter_opt: IterOption) -> Result<DBIterator<&DB>> {
-        let handle = self::util::get_cf_handle(self, cf)?;
-        let readopts = iter_opt.build_read_opts();
-        Ok(DBIterator::new_cf(self, handle, readopts))
-    }
-}
-
-impl Peekable for Snapshot {
-    fn get_value(&self, key: &[u8]) -> Result<Option<DBVector>> {
-        let mut opt = ReadOptions::new();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        let v = self.db.get_opt(key, &opt)?;
-        Ok(v)
-    }
-
-    fn get_value_cf(&self, cf: &str, key: &[u8]) -> Result<Option<DBVector>> {
-        let handle = self::util::get_cf_handle(&self.db, cf)?;
-        let mut opt = ReadOptions::new();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        let v = self.db.get_cf_opt(handle, key, &opt)?;
-        Ok(v)
-    }
-}
-
-impl Iterable for Snapshot {
-    fn new_iterator(&self, iter_opt: IterOption) -> DBIterator<&DB> {
-        let mut opt = iter_opt.build_read_opts();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        DBIterator::new(&self.db, opt)
-    }
-
-    fn new_iterator_cf(&self, cf: &str, iter_opt: IterOption) -> Result<DBIterator<&DB>> {
-        let handle = self::util::get_cf_handle(&self.db, cf)?;
-        let mut opt = iter_opt.build_read_opts();
-        unsafe {
-            opt.set_snapshot(&self.snap);
-        }
-        Ok(DBIterator::new_cf(&self.db, handle, opt))
-    }
-}
-
-impl Mutable for DB {}
-impl Mutable for WriteBatch {}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::rocks::Writable;
+    use super::Snapshot;
+    use crate::rocks::{util, Writable};
+    use crate::{Iterable, Mutable, Peekable};
     use kvproto::metapb::Region;
     use std::sync::Arc;
     use tempdir::TempDir;
@@ -195,15 +34,14 @@ mod tests {
     fn test_base() {
         let path = TempDir::new("var").unwrap();
         let cf = "cf";
-        let engine = Arc::new(
-            self::util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap(),
-        );
+        let engine =
+            Arc::new(util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap());
 
         let mut r = Region::new();
         r.set_id(10);
 
         let key = b"key";
-        let handle = self::util::get_cf_handle(&engine, cf).unwrap();
+        let handle = util::get_cf_handle(&engine, cf).unwrap();
         engine.put_msg(key, &r).unwrap();
         engine.put_msg_cf(handle, key, &r).unwrap();
 
@@ -233,8 +71,7 @@ mod tests {
     fn test_peekable() {
         let path = TempDir::new("var").unwrap();
         let cf = "cf";
-        let engine =
-            self::util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap();
+        let engine = util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap();
 
         engine.put(b"k1", b"v1").unwrap();
         let handle = engine.cf_handle("cf").unwrap();
@@ -249,9 +86,8 @@ mod tests {
     fn test_scan() {
         let path = TempDir::new("var").unwrap();
         let cf = "cf";
-        let engine = Arc::new(
-            self::util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap(),
-        );
+        let engine =
+            Arc::new(util::new_engine(path.path().to_str().unwrap(), None, &[cf], None).unwrap());
         let handle = engine.cf_handle(cf).unwrap();
 
         engine.put(b"a1", b"v1").unwrap();
